requirements.txt
Django
django-rest-framework
markdown
django-filter
